refactor: move client components to the leaves
